Implement player walk/run
Variables player speed with controller, with adjusted animations to match the different velocities.

Something strange is going on however. Walking/Running back&forth can cause Mike to bounce of the flat surface :/
Adding PhysicsMaterial2D without any bounciness doesn't solve the problem.

Experimented with setting Collider2D to Continuous checking, which reduced the amount of times it happens but does not remove it completely :/
Keeping it at Discrete for now...