Fixed bug where Jumping caused Mike to go immediately into falling state
There was a race condition between the moment player input is being processed, the state machine was being updated and the FixedUpdate() updating the statemachine during when no longer touching the ground.
Made sure that physics manipulations only happen during the FixedUpdate() calls.
And also prevented the 'JumpState' from being eligble for triggering a 'landing' and transitioning into any GroundedState.

This in turn also fixed another bug, that prevented the jump sound from playing when doing a Crouch Walk and then jumping.